+++ CHANGELOG.md
@@ -9,6 +9,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add the `share` and `unshare` events that are emitted when `share`/`unshare` methods are called.
+- Add the `open` and `close` events that are emitted when `open`/`close` methods are called.
+
+### Changed
+
+- **Breaking:** The `share`/`unshare` methods do not return a `MediaStream` anymore. The `managed` parameter now means that the stream is externally managed and should not be stopped automatically.
+
 ## [0.5.0] - 2026-06-25
 
 ### Added